requirements.txt
schedule==1.2.2
eventlet==0.37.0
python-dotenv==1.0.1
Flask-SocketIO==5.4.1